do not autoparse Message argument.
Automatically parsing into arg_list and arg_dict can cause
confusion about where to look for certain things.  For instance,
if you place a monitor on a structure, values will arrive in
arg_dict, but for a monitor on an integer, values will arrive in
arg_list[0].  Better to leave the argument alone and let user
call parse_argument() manually for obey parameters.