add debug output and fix numpy scalar conversion.
The 'obj[()]' syntax for converting a scalar array to a scalar
was hanging forever, possibly due to an odd threading setup
(main loop was in an imported module, so the main thread held
the import lock while drama ran in a separate thread).  To fix,
manually check shape and cast using dtype.type.